Remove BrowsableAPIRenderer and XMLRenderer
