--- taiga/base/api/static/api/css/bootstrap-tweaks.css
@@ -1,229 +1,0 @@
-/*
- * Copyright (C) 2014-2016 Andrey Antukh <niwi@niwi.nz>
- * Copyright (C) 2014-2016 Jesús Espino <jespinog@gmail.com>
- * Copyright (C) 2014-2016 David Barragán <bameda@dbarragan.com>
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- * The code is partially taken (and modified) from django rest framework
- * that is licensed under the following terms:
- *
- * Copyright (c) 2011-2014, Tom Christie
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * Redistributions of source code must retain the above copyright notice, this
- * list of conditions and the following disclaimer.
- * Redistributions in binary form must reproduce the above copyright notice, this
- * list of conditions and the following disclaimer in the documentation and/or
- * other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
- * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
- * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
- * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-/*
-
-This CSS file contains some tweaks specific to the included Bootstrap theme. 
-It's separate from `style.css` so that it can be easily overridden by replacing
-a single block in the template.
-
-*/
-
-
-.form-actions {
-    background: transparent;
-    border-top-color: transparent;
-    padding-top: 0;
-}
-
-.navbar-inverse .brand a {
-    color: #999;
-}
-.navbar-inverse .brand:hover a {
-    color: white;
-    text-decoration: none;
-}
-
-/* custom navigation styles */
-.wrapper .navbar{
-  width: 100%;
-  position: absolute;
-  left: 0;
-  top: 0;
-}
-
-.navbar .navbar-inner{
-  background: #2C2C2C;
-  color: white;
-  border: none;
-  border-top: 5px solid #A30000;
-  border-radius: 0px;
-}
-
-.navbar .navbar-inner .nav li, .navbar .navbar-inner .nav li a, .navbar .navbar-inner .brand:hover{
-  color: white;
-}
-
-.nav-list > .active > a, .nav-list > .active > a:hover {
-  background: #2c2c2c;
-}
-
-.navbar .navbar-inner .dropdown-menu li a, .navbar .navbar-inner .dropdown-menu li{
- color: #A30000;
-}
-.navbar .navbar-inner .dropdown-menu li a:hover{
-  background: #eeeeee;
-  color: #c20000;
-}
-
-/*=== dabapps bootstrap styles ====*/
-
-html{
-  width:100%;
-  background: none;
-}
-
-body, .navbar .navbar-inner .container-fluid {
-  max-width: 1150px;
-  margin: 0 auto;
-}
-
-body{
-  background: url("../img/grid.png") repeat-x;
-  background-attachment: fixed;
-}
-
-#content{
-    margin: 0;
-}
-
-/* sticky footer and footer */
-html, body {
-  height: 100%;
-}
-.wrapper {
-  min-height: 100%;
-  height: auto !important;
-  height: 100%;
-  margin: 0 auto -60px;
-}
-
-.form-switcher {
-    margin-bottom: 0;
-}
-
-.well {
-    -webkit-box-shadow: none;
-       -moz-box-shadow: none;
-            box-shadow: none;
-}
-
-.well .form-actions {
-    padding-bottom: 0;
-    margin-bottom: 0;
-}
-
-.well form {
-    margin-bottom: 0;
-}
-
-.well form .help-block {
-    color: #999;
-}
-
-.nav-tabs {
-    border: 0;
-}
-
-.nav-tabs > li {
-    float: right;
-}
-
-.nav-tabs li a {
-    margin-right: 0;
-}
-
-.nav-tabs > .active > a {
-    background: #f5f5f5;
-}
-
-.nav-tabs > .active > a:hover {
-    background: #f5f5f5;
-}
-
-.tabbable.first-tab-active .tab-content
-{
-    border-top-right-radius: 0;
-}
-
-#footer, #push {
-  height: 60px; /* .push must be the same height as .footer */
-}
-
-#footer{
-    text-align: right;
-}
-
-#footer p {
-  text-align: center;
-  color: gray;
-  border-top: 1px solid #DDD;
-  padding-top: 10px;
-}
-
-#footer a {
-  color: gray;
-  font-weight: bold;
-}
-
-#footer a:hover {
-  color: gray;
-}
-
-.page-header {
-    border-bottom: none;
-    padding-bottom: 0px;
-    margin-bottom: 20px;
-}
-
-/* custom general page styles */
-.hero-unit h2, .hero-unit h1{
-  color: #A30000;
-}
-
-body a, body a{
-  color: #A30000;
-}
-
-body a:hover{
-  color: #c20000;
-}
-
-#content a span{
-    text-decoration: underline;
- }
-
-.request-info {
-    clear:both;
-}
