--- taiga/base/api/static/api/css/default.css
@@ -1,114 +1,0 @@
-/*
- * Copyright (C) 2014-2016 Andrey Antukh <niwi@niwi.nz>
- * Copyright (C) 2014-2016 Jesús Espino <jespinog@gmail.com>
- * Copyright (C) 2014-2016 David Barragán <bameda@dbarragan.com>
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- * The code is partially taken (and modified) from django rest framework
- * that is licensed under the following terms:
- *
- * Copyright (c) 2011-2014, Tom Christie
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * Redistributions of source code must retain the above copyright notice, this
- * list of conditions and the following disclaimer.
- * Redistributions in binary form must reproduce the above copyright notice, this
- * list of conditions and the following disclaimer in the documentation and/or
- * other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
- * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
- * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
- * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-/* The navbar is fixed at >= 980px wide, so add padding to the body to prevent
-content running up underneath it. */
-
-h1 {
-    font-weight: 500;
-}
-
-h2, h3 {
-    font-weight: 300;
-}
-
-.resource-description, .response-info {
-    margin-bottom: 2em;
-}
-.version:before {
-    content: "v";
-    opacity: 0.6;
-    padding-right: 0.25em;
-}
-
-.version {
-  font-size: 70%;
-}
-
-.format-option {
-    font-family: Menlo, Consolas, "Andale Mono", "Lucida Console", monospace;
-}
-
-.button-form {
-    float: right;
-    margin-right: 1em;
-}
-
-ul.breadcrumb {
-  margin: 58px 0 0 0;
-}
-
-form select, form input, form textarea {
-  width: 90%;
-}
-
-form select[multiple] {
-  height: 150px;
-}
-/* To allow tooltips to work on disabled elements */
-.disabled-tooltip-shield {
-    position: absolute;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    left: 0;
-}
-
-.errorlist {
-    margin-top: 0.5em;
-}
-
-pre {
-  overflow: auto;
-  word-wrap: normal;
-  white-space: pre;
-  font-size: 12px;
-}
-
-.page-header {
-    border-bottom: none;
-    padding-bottom: 0px;
-    margin-bottom: 20px;
-}
-
