--- taiga/base/api/templates/api/api.html
@@ -1,3 +1,0 @@
-{% extends "api/base.html" %}
-
-{# Override this template in your own templates directory to customize #}
