--- taiga/base/api/templates/api/login.html
@@ -1,3 +1,0 @@
-{% extends "api/login_base.html" %}
-
-{# Override this template in your own templates directory to customize #}
