--- taiga/base/api/templates/api/raw_data_form.html
@@ -1,12 +1,0 @@
-{% load api %}
-{% csrf_token %}
-{{ form.non_field_errors }}
-{% for field in form %}
-    <div class="control-group">
-        {{ field.label_tag|add_class:"control-label" }}
-        <div class="controls">
-            {{ field }}
-            <span class="help-block">{{ field.help_text }}</span>
-        </div>
-    </div>
-{% endfor %}
